requirements.txt

# Main dependencies
PyYAML==6.0

# For creating standalone executables
pyinstaller==6.1.0

# Development dependencies
pytest==7.4.3
flake8==6.1.0
black==23.11.0
